requirements.txt
Flask==2.3.3
scikit-learn==1.3.0
numpy==1.24.3
pandas
PyPDF2==3.0.1
python-docx==0.8.11
nltk==3.8.1
joblib==1.3.2
xgboost==1.7.6
textblob==0.17.1
Werkzeug==2.3.7
pyresparser==1.0.6
spacy>=3.4.0
tensorflow>=2.10.0
keras>=2.10.0